basic terminal and minor changes
I've created a basic terminal that can run cd, ls and clear, I've also created the logic to be able to delete bytes from the screen, and modified the get_name function to return the file with a . between the extension and the name.

I also created a read_tree function that returns the values as a string, that will be treated however we like and made it so that any key press is processed before by the terminal, so we can see what we are writing, and modify it or change it anytime.